Add fix for non-breaking spaces in MacOS.
Adds a fix for a seriously frustrating problem.
The following post explains the cause and presents a fix:

https://relativkreativ.at/articles/why-chaining-commands-with-pipes-in-mac-os-x-does-not-always-work

I investigated the problem using the same techniques and came to the
same conclusion, but I had not yet come up with a fix until I came
across this blog post.

Seems to work well enough. Probably should investigate if this can
be used under Linux as well with no negative side effects. If not, the
inputrc file should be selectively sourced by some OS-detection scheme.